+++ CHANGELOG.md
@@ -21,7 +21,7 @@ way to update this template, but currently, we follow a pattern:
   [#960](https://github.com/sharetribe/flex-template-web/pull/960)
 * [fix] Remove unused translation keys and update PasswordChangePage title
   [#959](https://github.com/sharetribe/flex-template-web/pull/959)
-* [fix] Add translations CLI tool [#955](https://github.com/sharetribe/flex-template-web/pull/955)
+* [add] Add translations CLI tool [#955](https://github.com/sharetribe/flex-template-web/pull/955)
 
 ## [v2.3.2] 2018-11-20