+++ CHANGELOG.md
@@ -12,7 +12,9 @@ way to update this template, but currently, we follow a pattern:
 
 ---
 
-## Upcoming version 2018-11-XX
+## Upcoming version 2018-XX-XX
+
+## [v2.4.0] 2018-11-28
 
 * [change] Update Flex JS SDK to 1.2.
   [#963](https://github.com/sharetribe/flex-template-web/pull/963)
@@ -24,6 +26,8 @@ way to update this template, but currently, we follow a pattern:
 * [fix] Remove unused translation keys and update PasswordChangePage title
   [#959](https://github.com/sharetribe/flex-template-web/pull/959)
 * [add] Add translations CLI tool [#955](https://github.com/sharetribe/flex-template-web/pull/955)
+
+[v2.4.0]: https://github.com/sharetribe/flex-template-web/compare/v2.3.2...v2.4.0
 
 ## [v2.3.2] 2018-11-20